+++ main/serial.h
@@ -15,6 +15,6 @@ void serial_init(void);
 void serial_put_char(char);
 char serial_get_char(void);
 void serial_put_str(char *);
-void serial_put_int (uint32 num, int length);
+void serial_put_int (uint16 num, int length);
 
 #endif /* _SERIAL_H */
